Change paths internally and move shared to common
