+++ src/components/container/Container.stories.mdx
@@ -1,5 +1,5 @@
 import { Meta, Story, Preview, Props } from '@storybook/addon-docs/blocks';
-import { media, breakPoints } from '../shared/layout';
+import { media, breakPoints } from 'common/layout';
 import Container from './Container';
 
 <Meta title="Design System|layout/Media break points" component={Container} />
