+++ src/index.ts
@@ -3,12 +3,12 @@
  * importing them as `import { Component } from "package-name";`
  */
 
-import { colors } from './common/colors';
-import * as global from './components/shared/global';
-import { media, breakPoints } from './components/shared/layout';
+import { colors } from 'common/colors';
+import * as global from 'common/global';
+import { media, breakPoints } from 'common/layout';
 export { colors, global, media, breakPoints };
 
-export { default as Alert } from './components/alerts/Alert';
-export { default as TextField } from './components/forms/TextField';
-export { default as SpinnerLogo } from './components/logo/SpinnerLogo';
-export { default as StaticLogo } from './components/logo/StaticLogo';
+export { default as Alert } from 'components/alerts/Alert';
+export { default as TextField } from 'components/forms/TextField';
+export { default as SpinnerLogo } from 'components/logo/SpinnerLogo';
+export { default as StaticLogo } from 'components/logo/StaticLogo';
